+++ src/main/java/nhbank/core/domain/ADST_DPB_HISInfo.java
@@ -7,7 +7,8 @@ import javax.persistence.GeneratedValue;
 import javax.persistence.Id;
 import javax.persistence.Table;
 import java.io.Serializable;
-import java.sql.Date;
+import java.math.BigDecimal;
+import java.util.Date;
 
 @Entity
 @Table(name = "adstDpbHis")
@@ -16,15 +17,15 @@ public class ADST_DPB_HISInfo implements Serializable {
     @Id
     @GeneratedValue
     private String acctNo;
-    private int hisNo;
+    private BigDecimal hisNo;
     private String sts;
     private String cixNo;
     private String trGb;
     private String trCd;
     private String processCd;
     private String mechGb;
-    private int sdHisNo;
-    private int trAmt;
+    private BigDecimal sdHisNo;
+    private BigDecimal trAmt;
     private Date ibIl;
     private String trBr;
     private String sdBr;
@@ -34,10 +35,10 @@ public class ADST_DPB_HISInfo implements Serializable {
     private Date canIl;
     private String opNo;
     private String mngNo;
-    private int afJan;
+    private BigDecimal afJan;
     private Date acIl;
     private String chnlRefNo;
-    private int chnlHisNo;
+    private BigDecimal chnlHisNo;
     private String regEmpNo;
     private Date regDt;
     private String regTm;
